backend/requirements.txt
alembic
asyncpg
psycopg 
psycopg2-binary
sqlalchemy

# FastAPI
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings

redis

celery